% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surprisal_analysis.R
\name{plot_lambda}
\alias{plot_lambda}
\title{Plot lambda values over time}
\usage{
plot_lambda(lambda_values, lambda_no, time.points)
}
\arguments{
\item{lambda_values}{dataframe containing lambda values}

\item{lambda_no}{the pattern number of interest}

\item{time.points}{time points of the timecourse data}
}
\value{
a plot showing lambda values over time
}
\description{
Plot lambda values over time
}
