% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runSurprisalApp}
\alias{runSurprisalApp}
\title{Launch the SurprisalAnalysis Shiny App}
\usage{
runSurprisalApp(
  port = getOption("shiny.port", 3838),
  host = getOption("shiny.host", "127.0.0.1"),
  launch.browser = getOption("shiny.launch.browser", TRUE),
  run = TRUE,
  ...
)
}
\arguments{
\item{port}{port to run the app on (passed to shiny::runApp)}

\item{host}{host to listen on}

\item{launch.browser}{should launch a browser? set to TRUE by default}

\item{run}{boolean value, is set to TRUE by default. If set to FALSE it will not launch the graphical user interface}

\item{...}{Further arguments passed along to shiny::runApp}
}
\value{
no return value, running the function will launch an application with graphical user interface
}
\description{
Launch the SurprisalAnalysis Shiny App
}
\examples{
if (requireNamespace("httpuv", quietly = TRUE)) {
  runSurprisalApp(port = httpuv::randomPort(), run = FALSE)
}

}
