% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runSurprisalApp}
\alias{runSurprisalApp}
\title{Launch the SurprisalAnalysis Shiny App}
\usage{
runSurprisalApp(
  port = getOption("shiny.port", 3838),
  host = getOption("shiny.host", "127.0.0.1"),
  launch.browser = getOption("shiny.launch.browser", TRUE),
  ...
)
}
\arguments{
\item{port}{Integer; port to run the app on (passed to shiny::runApp)}

\item{host}{Character; host to listen on}

\item{launch.browser}{Logical; should launch a browser?}

\item{...}{Further args passed along to shiny::runApp}
}
\description{
Launch the SurprisalAnalysis Shiny App
}
