% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surprisal_analysis.R
\name{surprisal_analysis}
\alias{surprisal_analysis}
\title{This function performs surprisal analysis on transcriptomics data}
\usage{
surprisal_analysis(input.data)
}
\arguments{
\item{input.data}{transcriptomics data stores as dataframe}
}
\value{
a list containing the lambda values and corresponding weights of
transcripts stored
}
\description{
This function performs surprisal analysis on transcriptomics data
}
